Refactor tsembed package to be generic
